+++ NEWS.md
@@ -1,3 +1,8 @@
+## Changes in v1.4.0
+
+- Add `adjust_alpha` as an experimental argument to optimize `alpha` automatically.
+- Add `update_model` to update terms of existing models to classify documents with unseen words more accurately.
+
 ## Changes in v1.3.2
 
 - Improve the way to convert std::vector to arma::mat.